Refactoring FormAgepro Data-Interface methods to FormAgeproBase
